cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.20.0
